feat: split sidebar to panels
