+++ frontend/src/pages/AssetEditor/AssetEditor.tsx
@@ -12,8 +12,8 @@ import { useSearchParams } from 'react-router';
 import { toast } from 'react-toastify';
 
 import AssetMainProps from './AssetMainProps';
+import { AssetPreview } from './AssetPreview';
 import AssetEditorNav from './EditorNav';
-import Preview from './Preview';
 
 import nebula from '@/nebula';
 
@@ -475,7 +475,7 @@ const AssetEditor: React.FC<AssetEditorProps> = () => {
       case 'preview':
         return (
           <div className="grow row">
-            <Preview assetData={assetData} setAssetData={setAssetData} />
+            <AssetPreview assetData={assetData} setAssetData={setAssetData} />
           </div>
         );
