+++ frontend/src/pages/AssetEditor/AssetPreview/index.ts
@@ -1,0 +1,1 @@
+export { AssetPreview } from './AssetPreview';
